small planing fixes

+++ BO/Models/PlanitPoker/DAL/PlaningStoryDal.cs
@@ -8,7 +8,7 @@ namespace BO.Models.PlaningPoker.DAL
         public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Vote { get; set; }
+        public decimal Vote { get; set; }
         public DateTime Date { get; set; }
         public bool Completed { get; set; }
